exercicio 6

+++ 05-exercicio-parte-2.rb
@@ -3,7 +3,7 @@ nome = gets.chomp
 
 
 print "Agora informe sua idade: "
-idade = gets.to_f
+idade = gets.to_i
 
 def calcula_estimativa_batimentos(idade)
     return idade * 365 * 24 * 60 * 80
